requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Authlib==1.2.1
boto3==1.26.49
botocore==1.29.49
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.3
cryptography==41.0.4
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1.5
django-extensions==3.2.1
django-heroku==0.3.1
dnspython==2.4.2
et-xmlfile==1.1.0
filelock==3.9.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
joblib==1.2.0
loguru==0.7.2
multidict==6.0.4
nltk==3.8.1
numpy==1.26.1
openai==0.26.1
openpyxl==3.0.10
pandas==2.1.1
pandas-stubs==1.5.2.230105
pinecone-client==2.2.4
platformdirs==2.6.0
psycopg2==2.9.9
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2022.10.31
requests==2.31.0
s3transfer==0.6.0
six==1.16.0
smart-open==6.3.0
sqlparse==0.4.3
tqdm==4.66.1
types-pytz==2022.7.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
validators==0.22.0
virtualenv==20.17.1
weaviate-client==3.24.2
whitenoise==6.3.0
yarl==1.8.2
